// src/api-chain/model/apiResponsePaginatedListChainPayment.ts
export interface ApiResponsePaginatedListChainPayment { 
    data?: PaginatedListChainPayment;
    /**
     * Optional details for unexpected error responses.
     */
    errorDetails?: string;
    /**
     * Simple message to explain client developers the reason for error.
     */
    errorMessage?: string;
    status: ApiDefaultResponseStatusEnum;
    validationErrorDetails?: ApiValidationErrorDetails;
}

/**
 * Namespace for property- and property-value-enumerations of ApiResponsePaginatedListChainPayment.
 */
export namespace ApiResponsePaginatedListChainPayment {
    /**
     * All properties of ApiResponsePaginatedListChainPayment.
     */
    export enum Properties {
        data = 'data',
        /**
         * Optional details for unexpected error responses.
         */
        errorDetails = 'errorDetails',
        /**
         * Simple message to explain client developers the reason for error.
         */
        errorMessage = 'errorMessage',
        status = 'status',
        validationErrorDetails = 'validationErrorDetails'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiResponsePaginatedListChainPayment',
            vars: [
                {
                    metadata: PaginatedListChainPayment.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'data',
                    classname: 'ApiResponsePaginatedListChainPayment',
                    dataType: 'PaginatedListChainPayment',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'PaginatedListChainPayment'
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'errorDetails',
                    classname: 'ApiResponsePaginatedListChainPayment',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'errorMessage',
                    classname: 'ApiResponsePaginatedListChainPayment',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
                {
                    metadata: ApiDefaultResponseStatusEnum.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: true,
                    name: 'status',
                    classname: 'ApiResponsePaginatedListChainPayment',
                    dataType: 'ApiDefaultResponseStatusEnum',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ApiDefaultResponseStatusEnum'
                },
                {
                    metadata: ApiValidationErrorDetails.formMetadata,
                    isReadOnly: false,
                    isEnum: false,
                    required: false,
                    name: 'validationErrorDetails',
                    classname: 'ApiResponsePaginatedListChainPayment',
                    dataType: 'ApiValidationErrorDetails',
                    isPrimitiveType: false,
                    isListContainer: false,
                    complexType: 'ApiValidationErrorDetails'
                },
            ],
            validators: {
                data: [
                ],
                errorDetails: [
                ],
                errorMessage: [
                ],
                status: [
                        ['required'],
                ],
                validationErrorDetails: [
                ],
            }
        }
    }

  // export const ApiResponsePaginatedListChainPaymentValidationScheme = {
  //     validators: [],
  //     fields: {
  //               data: {
  //                   validators: []
  //               },
  //               errorDetails: {
  //                   validators: []
  //               },
  //               errorMessage: {
  //                   validators: []
  //               },
  //               status: {
  //                   validators: []
  //               },
  //               validationErrorDetails: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiResponsePaginatedListChainPayment>;


}